question1/requirements.txt
Biopython==1.73
